Migrate Pinned Messages room action
