+++ apps/meteor/client/startup/contextualBar/index.ts
@@ -1,2 +1,1 @@
-import './pinnedMessagesTab';
 import './starredMessagesTab';
